Print commands as they are executing
The rationale behind this change is that it makes it obvious to the
developers to see what is actually happening in case a build fails.
Prior to this change, it was clear that e.g. a project was
configuring/building/running tests, but not really what the specific
command was.