Install straight to the target directory
Some applications/libraries such as kio do not cope well with being
relocated without being told. In case of KIO, the path to the `kioslave`
binary was hardcoded. This was never a problem on build.kde.org because
builds are being deployed all the time there, but it is a problem on a
try-without-deply CI.