Prepare for Qt5 installation through the CI system
The Qt5 versions which we're going to use start at 5.2 (the oldest one
still supported by KF5). The release builds are as a rule of thumb
faster, so this is what we prefer for the older releases. The most
recent Qt version, however, is configured in the debug mode to ensure
that we benefit from the extra asserts and what not, especially when
running our unit tests.

- 5.2 release mode
- 5.3 release mode
- 5.4 debug build