Enable simple symlinking of some heavy build deps
This is motivated by a situation where a build node trashes its entire
installation directory between jobs. As e.g. the debug build of Qt 5.4
consumes 2.8 GB on disk *and* is not expected to change at all once
created (or can be easily rebaked for the VM image if it ever changes),
it doesn't make sense to transfer it during each build job. The wasted
time adds up pretty quickly.