Push systemwide entries for XDG_* vars to the very end
The script was adding the usual system-wide paths while processing each
dependency data from the current project's list of dependencies, which
led to many instances of /usra/share (etc) on places which were "too
early" in the search path.

This was causing real breakages for e.g. kio's unit tests. Thanks to
David Faure <dfaure@kde.org> for catching this, investigating this, and
identifying a problem in the CI infrastructure.